functions/get-currency-rates/requirements.txt
google-cloud-bigquery==2.13.0
Flask==1.0.2